osbuild-service-maintenance: Move maintenance queries out of jobqueue
